vspreview/main.py: change -a to follow vspipe convention of passing via global variables
Passing argument via global variables is generally better than re-interpreting the
string as shell command line via shlex, which necessitates double escaping. Not to
mention that this makes the vpy script behave the same when running under vspipe
and vs-preview.

The only difference between this change and vspipe's -a is that we're passing the
value as Python str instead of bytes, which seems more natural.

Signed-off-by: akarin <i@akarin.info>