Fix frames props view not updating on frame change
2818451 updated the plugin to use a single `TableModel` instance, but
changes to the underlying data are not emitted as signals, so the view
doesn't update when the data changes. This patch fixes the issue by
emitting the `layoutChanged` signal to trigger a redisplay.

Didn't use `dataChanged` because it's possible that there are more or
less rows (props) in the new table data, hence a layout change. Also
didn't implement this in `TableModel` since the data mutation is
happening directly in the plugin.