Add fallback constructor for unknown python/object tags
This prevents storage loading from erroring out when loading an old
storage that references classes that no longer exist. (vspreview may
still error out later when it gets None in places where it doesn't
expect it, but that can then be handled with additional checks)

This fixes one of the two causes of crashes due to outdated storage
files (the other one being aliases in the local storage pointing to
removed fields in the global storage).