Wrap YAML Loader to not error out on broken aliases
This fixes the second set of crashes due to outdated config files.
One unfortunate consequence is that we can no longer use the C loaders,
since we need to override methods. If this has a noticable performance
impact, one could first try loading using the C loader and then fall
back to the loader wrapper in the case of errors.

Also, some of this code no longer feels like it belongs in bases.py but
oh well.