Make YAML storage loading safer and able to catch errors (#206)
* Switch to safe YAML loader

This prevents arbitrary code execution from access to config files
using tags like !!python/object/apply:os.system and in the process
sets up some infrastructure for eventually switching to storage using
custom tags that does not use python/object.
It does, however, need a bit of yaml constructor spaghetti to parse
some of the datatypes that can't simply inherit from SafeYAMLObject.

* Add fallback constructor for unknown python/object tags

This prevents storage loading from erroring out when loading an old
storage that references classes that no longer exist. (vspreview may
still error out later when it gets None in places where it doesn't
expect it, but that can then be handled with additional checks)

This fixes one of the two causes of crashes due to outdated storage
files (the other one being aliases in the local storage pointing to
removed fields in the global storage).

* Wrap YAML Loader to not error out on broken aliases

This fixes the second set of crashes due to outdated config files.
One unfortunate consequence is that we can no longer use the C loaders,
since we need to override methods. If this has a noticable performance
impact, one could first try loading using the C loader and then fall
back to the loader wrapper in the case of errors.

Also, some of this code no longer feels like it belongs in bases.py but
oh well.